example changes and move internal folder changes to util
