+++ memory/cloudsql/chat_message_history.go
@@ -5,10 +5,10 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
+	"github.com/tmc/langchaingo/util/cloudsqlutil"
 	"time"
 
 	"github.com/jackc/pgx/v5"
-	"github.com/tmc/langchaingo/internal/cloudsqlutil"
 	"github.com/tmc/langchaingo/llms"
 	"github.com/tmc/langchaingo/schema"
 )
