+++ vectorstores/alloydb/vectorestores.go
@@ -4,13 +4,13 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
+	"github.com/tmc/langchaingo/util/alloydbutil"
 	"strconv"
 	"strings"
 
 	"github.com/google/uuid"
 	"github.com/jackc/pgx/v5"
 	"github.com/tmc/langchaingo/embeddings"
-	"github.com/tmc/langchaingo/internal/alloydbutil"
 	"github.com/tmc/langchaingo/schema"
 	"github.com/tmc/langchaingo/vectorstores"
 )
@@ -79,7 +79,7 @@ func (vs *VectorStore) AddDocuments(ctx context.Context, docs []schema.Document,
 		}
 	}
 	// If no metadata provided, initialize with empty maps
-	metadatas := make([]map[string]any, len(texts))
+	metadatas := make([]map[string]any, len(docs))
 	for i := range docs {
 		if docs[i].Metadata == nil {
 			metadatas[i] = make(map[string]any)
@@ -94,7 +94,6 @@ func (vs *VectorStore) AddDocuments(ctx context.Context, docs []schema.Document,
 		content := texts[i]
 		embedding := vectorToString(embeddings[i])
 		metadata := metadatas[i]
-
 		// Construct metadata column names if present
 		metadataColNames := ""
 		if len(vs.metadataColumns) > 0 {
@@ -115,7 +114,7 @@ func (vs *VectorStore) AddDocuments(ctx context.Context, docs []schema.Document,
 			if val, ok := metadata[metadataColumn]; ok {
 				valuesStmt += fmt.Sprintf(", $%d", len(values)+1)
 				values = append(values, val)
-				delete(metadata, metadataColumn)
+				// delete(metadata, metadataColumn)
 			} else {
 				valuesStmt += ", NULL"
 			}
@@ -132,6 +131,7 @@ func (vs *VectorStore) AddDocuments(ctx context.Context, docs []schema.Document,
 		valuesStmt += ")"
 		query := insertStmt + valuesStmt
 		b.Queue(query, values...)
+
 	}
 
 	batchResults := vs.engine.Pool.SendBatch(ctx, b)
