+++ vectorstores/cloudsql/vectorstore_options.go
@@ -3,7 +3,7 @@ package cloudsql
 import (
 	"errors"
 	"github.com/tmc/langchaingo/embeddings"
-	"github.com/tmc/langchaingo/internal/cloudsqlutil"
+	"github.com/tmc/langchaingo/util/cloudsqlutil"
 	"github.com/tmc/langchaingo/vectorstores"
 )
